GCP/CloudFunctions/trigger-photo/requirements.txt
functions-framework==3.*
psycopg2 == 2.9.9
google-cloud-aiplatform==1.33.1
google-cloud-storage==2.11.0
Pillow==10.0.1
numpy==1.26.0
Python-IO==0.3
opencv-python==4.8.0.74